Grasshopper Components for the new dynamic relaxation engine in the BHoM Engine
The structural DR is kind of similar to kangaroo2 with goals and stuff.
But there are some more things that needs to be adjusted manually.
So far all the goals that exists is a unary force, a spring and a
special one for cable net form finding and prestress finding.
The old Dynamic Relaxation is put in a superseeded folder.